[PickerTogglerTag]: removed tooltipContext and made a doc example
