+++ app/src/sandbox/SandboxPage.tsx
@@ -23,7 +23,6 @@ import { ProjectTasksDemo } from './tasks/ProjectTasksDemo';
 import { RichTextEditorDemo } from './RTE/rteDemo';
 import { TableColumnConfigModalTest } from './tableColConfigModal/TableColumnConfigModalTest';
 import { PalettePage } from './tokens/palette/palettePage';
-import { PickerTogglerTagDemo } from './pickerTogglerTag/PickerTogglerTagDemo';
 
 export function SandboxPage() {
     const items = useMemo(
@@ -47,7 +46,6 @@ export function SandboxPage() {
             { id: 'tagInput', name: 'Tag Input', component: TagInputDemo },
             { id: 'tokens', name: 'Tokens' },
             { parentId: 'tokens', id: 'tokensPalette', name: 'Palette', component: PalettePage },
-            { id: 'pickerTogglerTag', name: 'PickerTogglerTag', component: PickerTogglerTagDemo },
         ],
         [],
     );
