+++ public/docs/content/examples-pickerInput-PickerTogglerTagDemo.json
@@ -1,0 +1,17 @@
+[
+  {
+    "type": "paragraph",
+    "children": [
+      {
+        "text": "In this example we use "
+      },
+      {
+        "text": "renderTag",
+        "uui-richTextEditor-code": true
+      },
+      {
+        "text": " callback to make custom Tags with Tooltip if they are not a Tag for collapsed values."
+      }
+    ]
+  }
+]
