+++ uui/components/pickers/PickerTogglerTag.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import * as types from '../types';
 import { Tag, TagProps } from '../widgets';
-import { Tooltip } from '../overlays';
 import { DataRowProps } from '@epam/uui-core';
 
 export type PickerTogglerTagProps<TItem, TId> = TagProps & {
@@ -9,8 +8,6 @@ export type PickerTogglerTagProps<TItem, TId> = TagProps & {
     size?: types.ControlSize;
     /** If this is true, then the PickerTogglerTag will be an additional tag with the number of collapsed elements in the caption. */
     isCollapsed?: boolean;
-    /** Defines content for tooltip */
-    tooltipContent?: string;
     /** Defines row props (see more: uui-components/src/pickers/PickerToggler.tsx PickerTogglerProps<TItem = any, TId = any>) */
     rowProps?: DataRowProps<TItem, TId>;
 };
@@ -32,18 +29,12 @@ const getPickerTogglerButtonSize = (propSize?: types.ControlSize):TagProps['size
     }
 };
 
-export function PickerTogglerTag(props: PickerTogglerTagProps<any, any>): JSX.Element {
+export const PickerTogglerTag = React.forwardRef((props: PickerTogglerTagProps<any, any>, ref: React.Ref<HTMLSpanElement>): JSX.Element => {
     const tagProps = {
         ...props,
         tabIndex: -1,
         size: getPickerTogglerButtonSize(props.size),
     };
 
-    return (props.tooltipContent)
-        ? (
-            <Tooltip content={ props.tooltipContent } openDelay={ 400 }>
-                <Tag { ...tagProps } />
-            </Tooltip>
-        )
-        : <Tag { ...tagProps } />;
-}
+    return <Tag ref={ ref } { ...tagProps } />;
+});
